refactor(product): split desktop file actions
